fix: 3 bugs — qte decimal, caisse-statut polling, variant deactivation on delete
- #49: remove .int() from vente qte schema — allows decimal kg quantities
- #57: add refetchInterval: 30_000 on caisse-statut query in POS page
- #52: variant trash with DB id now deactivates in DB instead of silently orphaning

+++ apps/api/src/routes/ventes.ts
@@ -14,7 +14,7 @@ const venteCreateSchema = z.object({
     .array(
       z.object({
         varianteProduitId: z.string().min(1),
-        qte: z.number().int().positive(),
+        qte: z.number().positive(),
         remise: z.number().min(0).max(100).optional(),
       }),
     )
